model: Override all font sizes
Copies a bunch of rules from the default theme and the Thematic theme,
adjusting their font sizes.

Also adds a Web CSS override, to control the font sizes inside the
article. (This doesn't work yet for the Dinosaurs app because of an
omission in the API: https://phabricator.endlessm.com/T23560)